+++ html/js/config/index.js
@@ -3,6 +3,7 @@ import { configBackgroundImages } from './background-images.js';
 import { configBackgroundMusic } from './background-music.js';
 import { configBackgroundVideos } from './background-videos.js';
 import { configErrorLog, setupErrorLog } from './error-log.js';
+import { configGenerator } from './generator.js';
 import { configMessages } from './messages.js';
 import { configStyle, setupStyle } from './style.js';
 
@@ -23,4 +24,5 @@ export function config(handoverData) {
     configBackgroundVideos(handoverData);
     configBackgroundMusic(handoverData);
     configAudioControls(handoverData);
+    configGenerator(handoverData);
 }